refactor: extract EmojiPicker component
