+++ packages/client/src/components/EmojiReactions/index.ts
@@ -1,1 +1,2 @@
 export { default as EmojiReactions } from './EmojiReactions'
+export { default as EmojiPicker } from './EmojiPicker'
